linting and stuff
